backend/requirements.txt
asgiref==3.9.1
Django==5.2.6
django-cors-headers==4.8.0
djangorestframework==3.16.1
huepy==1.2.1
Jinja2==3.1.6
livereload==2.7.1
markdown2==2.5.4
MarkupSafe==3.0.2
Pygments==2.19.2
sqlparse==0.5.3
tornado==6.5.2
